refactor: separate context and provider to improve fast refresh
